move checkParams to utils
